# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WTR Ghana Water Platform API
  version: 1.0.0
  description: Multi-tenant community water utility management API.
paths:
...
  /api/customers/{id}/:
    get:
      operationId: customers_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      tags:
      - customers
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: ''
    put:
      operationId: customers_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      tags:
      - customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Customer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: ''
    patch:
      operationId: customers_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      tags:
      - customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: ''
    delete:
      operationId: customers_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      tags:
      - customers
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    AccountStatusEnum:
      enum:
      - PENDING
      - ACTIVE
      - SUSPENDED
      - DISCONNECTED
      - INACTIVE
      type: string
      description: |-
        * `PENDING` - Pending connection
        * `ACTIVE` - Active
        * `SUSPENDED` - Suspended
        * `DISCONNECTED` - Disconnected
        * `INACTIVE` - Inactive
...
    CategoryDc2Enum:
      enum:
      - RESIDENTIAL
      - COMMERCIAL
      - INSTITUTIONAL
      - INDUSTRIAL
      - SCHOOL
      - CHURCH
      - MOSQUE
      - PUBLIC_FACILITY
      - GOVERNMENT
      type: string
      description: |-
        * `RESIDENTIAL` - Residential
        * `COMMERCIAL` - Commercial
        * `INSTITUTIONAL` - Institutional
        * `INDUSTRIAL` - Industrial
        * `SCHOOL` - School
        * `CHURCH` - Church
        * `MOSQUE` - Mosque
        * `PUBLIC_FACILITY` - Public facility
        * `GOVERNMENT` - Government facility
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        town_name:
          type: string
          readOnly: true
        community_name:
          type: string
          readOnly: true
        property_label:
          type: string
          readOnly: true
        tariff_plan_name:
          type: string
          readOnly: true
        meter:
          type: string
          readOnly: true
        portal_email:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        customer_id:
          type: string
          readOnly: true
        household_name:
          type: string
          maxLength: 150
        contact_person:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 30
        alternative_phone:
          type: string
          maxLength: 30
        email:
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        address:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        category:
          $ref: '#/components/schemas/CategoryDc2Enum'
        occupants:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        connection_date:
          type: string
          format: date
          nullable: true
        account_status:
          $ref: '#/components/schemas/AccountStatusEnum'
        outstanding_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        risk_score:
          type: integer
          readOnly: true
        risk_level:
          type: string
          readOnly: true
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        town:
          type: integer
          nullable: true
        property:
          type: integer
          nullable: true
        tariff_plan:
          type: integer
          nullable: true
        user:
          type: integer
          readOnly: true
          nullable: true
      required:
      - community
      - community_name
      - contact_person
      - created_at
      - customer_id
      - household_name
      - id
      - meter
      - outstanding_balance
      - phone
      - portal_email
      - property_label
      - risk_level
      - risk_score
      - tariff_plan_name
      - town_name
      - updated_at
      - user
...
    PatchedCustomer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        town_name:
          type: string
          readOnly: true
        community_name:
          type: string
          readOnly: true
        property_label:
          type: string
          readOnly: true
        tariff_plan_name:
          type: string
          readOnly: true
        meter:
          type: string
          readOnly: true
        portal_email:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        customer_id:
          type: string
          readOnly: true
        household_name:
          type: string
          maxLength: 150
        contact_person:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 30
        alternative_phone:
          type: string
          maxLength: 30
        email:
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        address:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        category:
          $ref: '#/components/schemas/CategoryDc2Enum'
        occupants:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        connection_date:
          type: string
          format: date
          nullable: true
        account_status:
          $ref: '#/components/schemas/AccountStatusEnum'
        outstanding_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        risk_score:
          type: integer
          readOnly: true
        risk_level:
          type: string
          readOnly: true
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        town:
          type: integer
          nullable: true
        property:
          type: integer
          nullable: true
        tariff_plan:
          type: integer
          nullable: true
        user:
          type: integer
          readOnly: true
          nullable: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT